Update branch to latest changes on #10
