+++ requirements.txt
@@ -3,3 +3,4 @@ pandas==1.4.2
 nltk==3.6.1
 scipy==1.8.1
 scikit-learn==1.1.1
+joblib==1.1.0
