+++ src/evaluate.py
@@ -1,17 +1,27 @@
+from typing import Dict
+
+import numpy as np
 import json
-from joblib import load
+
 from sklearn.metrics import accuracy_score
-from sklearn.metrics import average_precision_score
 from sklearn.metrics import f1_score
+from sklearn.metrics import average_precision_score
+
+from sklearn.base import BaseEstimator
+
+from joblib import load
 
 
-def print_evaluation_scores(y_val, predicted):
-    print('Accuracy score: ', accuracy_score(y_val, predicted))
-    print('F1 score: ', f1_score(y_val, predicted, average='weighted'))
-    print('Average precision score: ', average_precision_score(y_val, predicted, average='macro'))
+def print_evaluation_scores(y_val: np.ndarray, predicted: np.ndarray):
+    print("Accuracy score: ", accuracy_score(y_val, predicted))
+    print("F1 score: ", f1_score(y_val, predicted, average="weighted"))
+    print(
+        "Average precision score: ",
+        average_precision_score(y_val, predicted, average="macro"),
+    )
 
 
-def save_evaluation_scores(y_val, predicted, algorithm):
+def save_evaluation_scores(y_val: np.ndarray, predicted: np.ndarray, algorithm: str):
     res = {
         'Accuracy': accuracy_score(y_val, predicted),
         'F1 score': f1_score(y_val, predicted, average='weighted'),
@@ -21,33 +31,44 @@ def save_evaluation_scores(y_val, predicted, algorithm):
         json.dump(res, f)
 
 
-def print_words_for_tag(classifier, tag, tags_classes, index_to_words):
+def print_words_for_tag(
+        classifier: BaseEstimator,
+        tag: str,
+        tags_classes: list,
+        index_to_words: Dict[int, str],
+):
     """
-        Print top 5 positive and top 5 negative words for current tag
-        Parameters
-        ---------
-        classifier
-                trained classifier
-        tag
-                a particular tag
-        tags_classes
-                list of classes names from MultiLabelBinarizer
-        index_to_words
-                index_to_words transformation
-        all_words
-                all words in the dictionary
+    Print top 5 positive and top 5 negative words for current tag
+
+    Parameters
+    ---------
+    classifier
+            trained classifier
+    tag
+            a particular tag
+    tags_classes
+            list of classes names from MultiLabelBinarizer
+    index_to_words
+            index_to_words transformation
+    all_words
+            all words in the dictionary
     """
-    print('Tag:\t{}'.format(tag))
+
+    print("fTag:\t{tag}")
 
     # Extract an estimator from the classifier for the given tag.
     # Extract feature coefficients from the estimator.
 
     model = classifier.estimators_[tags_classes.index(tag)]
-    top_positive_words = [index_to_words[x] for x in model.coef_.argsort().tolist()[0][-5:]]
-    top_negative_words = [index_to_words[x] for x in model.coef_.argsort().tolist()[0][:5]]
+    top_positive_words = [
+        index_to_words[x] for x in model.coef_.argsort().tolist()[0][-5:]
+    ]
+    top_negative_words = [
+        index_to_words[x] for x in model.coef_.argsort().tolist()[0][:5]
+    ]
 
-    print('Top positive words:\t{}'.format(', '.join(top_positive_words)))
-    print('Top negative words:\t{}\n'.format(', '.join(top_negative_words)))
+    print(f"Top positive words:\t{', '.join(top_positive_words)}")
+    print(f"Top negative words:\t{', '.join(top_negative_words)}\n")
 
 
 def main():
