+++ tests/test_bag_of_words.py
